[PVR] Cleanup: Move CPVRClientCapabiliities to its own set of source files.
