+++ xbmc/pvr/addons/CMakeLists.txt
@@ -1,9 +1,11 @@
 set(SOURCES PVRClient.cpp
+            PVRClientCapabilities.cpp
             PVRClientMenuHooks.cpp
             PVRClientUID.cpp
             PVRClients.cpp)
 
 set(HEADERS PVRClient.h
+            PVRClientCapabilities.h
             PVRClientMenuHooks.h
             PVRClientUID.h
             PVRClients.h)
